First attempt implementing parseSetExpression in expressionParser
Should effectively lay the groundwork for implementing sets. It will
handle any set without a variable in the range; this is because I made
the false assumption that the symbol table had values at this stage (it
only has keys). Probably need to create a "RANGE" node with lower and
upper bound children. Also for some reason even though it parses the
sets correctly, the rootNode is null when returned.