Set parsing complete! (I hope)
Added DUPLICATE_SET_VALUE error.
Fixed last element of set not being added to parse tree.
Fixed off by one in INTEGER_CONSTANT range evaluation.

INTEGER_CONSTANTs are no longer stored as nodes in the parse tree;
anything that can be evaluated is put into the SET_VALUES node's
HashSet<Integer> for easy lookup. At runtime the remaining nodes should
be evaluated and then their values added to the HashSet for efficient
lookups. Evaluated nodes can be left and ignored, or deleted if you are
in the mood for destruction.